Fixed test data

+++ src/test/java/dto/api/user/rq/User.java
@@ -31,5 +31,5 @@ public class User {
     private String sex;
     @SerializedName("money")
     @Expose
-    private long money;
+    private Double money;
 }
